# Improved suffix benchmark performance